+++ doc/source/changelog.md
@@ -17,6 +17,7 @@ Semver is not yet in place, so each version can have breaking changes, although 
     - Added support for block generation on the GPU (only available with generators that support both CPU and GPU, for now `VoxelGeneratorGraph` only).
     - Updated FastNoise2 to 0.10.0-alpha
     - Started an experimental type system for the blocky voxels workflow. However it is not fully functional, its API may change in the future or have parts removed.
+    - Added experimental `VoxelAStarGrid3D` for grid-based pathfinding on blocky voxels 
     - Voxel engine processing no longer stops when the SceneTree is paused
     - `VoxelGeneratorGraph`:
         - Added `Spots2D` and `Spots3D` nodes, optimized for generating "ore patches"